re-add old change_configmaps.js, rename new configmap script
